pyproject.toml
[project]
name = "api"
version = "1.0.0"
requires-python = ">=3.14"
dependencies = []

[dependency-groups]
dev = ["ruff", "pytest", "pytest-xdist", "pytest-cov", "ty"]
# Shared base dependencies
base = ["boto3>=1.34.16"]
db = [{include-group = "base"}, "pynamodb>=5.5.1"]
http = [{include-group = "base"}, "requests>=2.28.1"]
billing = [{include-group = "db"}, "stripe>=5.0.0"]
# Lambda-specific dependency groups
trade = [{include-group = "http"}, "robin-stocks>=3.4.0", "python-jose>=3.3.0"]
account = [{include-group = "billing"}]
contact = [{include-group = "db"}]
gym = [{include-group = "db"}, {include-group = "http"}, "pandas>=2.2.0"]
model = [{include-group = "base"}, "numpy>=1.26.0"]
notify = [{include-group = "db"}, {include-group = "http"}, "jinja2>=3.1.2"]
preview = [{include-group = "db"}]
signals = [{include-group = "db"}]
subscribe = [{include-group = "billing"}]

[tool.pytest.ini_options]
# Parallel execution disabled - tests have sequential dependencies
# addopts = "-n auto"
testpaths = ["tests/api"]
pythonpath = ["src/api", "src/api/shared/python"]

[tool.coverage.run]
source = ["src/api"]
omit = ["**/test/*", "**/tests/*", ".venv/*", "src/api/trade/app.py", "src/api/shared/python/auth.py"]

[tool.coverage.report]
fail_under = 85 # increase this to 95%
show_missing = true

[tool.ruff]
target-version = "py314"
line-length = 88
exclude = [
    ".aws-sam",
    ".venv",
    "*.ipynb",
    "test.py",
    "fix_account.py", 
    "send_email.py",
    "chance_of_profit.py",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations (type hints)
    "D",      # pydocstyle (docstrings)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "ANN401", # dynamically typed expressions (Any)
    "D100",   # missing docstring in public module
    "D104",   # missing docstring in public package
    "D203",   # 1 blank line required before class docstring (conflicts with D211)
    "D213",   # multi-line docstring summary should start at second line (conflicts with D212)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/test/*" = ["ANN", "D", "F405", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty]
[tool.ty.environment]
python-version = "3.14"
extra-paths = ["src/api/shared/python"]

[tool.ty.rules]
# Ignore third-party import errors since lambdas have different dependency groups
unresolved-import = "ignore"
